+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/ServerPools/IServerPool.cs
@@ -41,6 +41,9 @@ namespace Stormancer.Server.Plugins.GameSession.ServerPool
         public DateTime CreatedOn { get; internal set; }
         public IScenePeerClient Peer { get; internal set; }
         public TaskCompletionSource<GameServerStartupParameters> RunTcs { get; internal set; }
+        public object? Context { get; internal set; }
+        public GameSessionConfiguration GameSessionConfiguration { get; internal set; }
+        public TaskCompletionSource<WaitGameServerResult> RequestCompletedCompletionSource { get; internal set; }
 
         public void Dispose()
         {
@@ -58,6 +61,14 @@ namespace Stormancer.Server.Plugins.GameSession.ServerPool
         public SessionId GameServerSessionId { get; set; }
     }
 
+    public class WaitGameServerResult
+    {
+        [MemberNotNullWhen(true,"Value")]
+        public bool Success { get; set; }
+
+      
+        public GameServer? Value { get; set; }
+    }
     
     public interface IServerPool: IDisposable
     {
@@ -69,7 +80,7 @@ namespace Stormancer.Server.Plugins.GameSession.ServerPool
         /// <param name="gameSessionId"></param>
         /// <param name="gameSessionConfig"></param>
         /// <returns></returns>
-        Task<GameServer> WaitGameServerAsync(string gameSessionId, GameSessionConfiguration gameSessionConfig, CancellationToken cancellationToken);
+        Task<WaitGameServerResult> TryWaitGameServerAsync(string gameSessionId, GameSessionConfiguration gameSessionConfig, CancellationToken cancellationToken);
 
         void UpdateConfiguration(JObject config);