+++ src/Stormancer.Plugins/GameVersion/Stormancer.Server.Plugins.GameVersion/AuthenticationEventHandler.cs
@@ -58,6 +58,11 @@ namespace Stormancer.Server.Plugins.GameVersion
             {
                 return Task.CompletedTask;
             }
+            //Don't control the gameversion on "services" clients. Services clients include gameserver hosting agents.
+            if(authenticationCtx.AuthCtx.CurrentSession.SessionData.TryGetValue("stormancer.type", out var bytes) && System.Text.Encoding.UTF8.GetString(bytes) == "service")
+            {
+                return Task.CompletedTask;
+            }
 
             var serverVersion = gameVersionService.CurrentGameVersion;