+++ src/Stormancer.Plugins/Profile/Stormancer.Server.Plugins.Profile/ProfileController.cs
@@ -50,24 +50,24 @@ namespace Stormancer.Server.Plugins.Profile
         }
 
         [Api(ApiAccess.Public, ApiType.Rpc)]
-        public async Task<Dictionary<string, ProfileDto>> GetProfilesBySessionIds(IEnumerable<string> sessionIds, Dictionary<string, string> displayOptions, RequestContext<IScenePeerClient> ctx)
+        public async Task<Dictionary<SessionId, ProfileDto>> GetProfilesBySessionIds(IEnumerable<SessionId> sessionIds, Dictionary<string, string> displayOptions, RequestContext<IScenePeerClient> ctx)
         {
 
-            var sessions = await _sessions.GetSessions(sessionIds.Select(s=>SessionId.From(s)), ctx.CancellationToken);
+            var sessions = await _sessions.GetSessions(sessionIds, ctx.CancellationToken);
 
             IEnumerable<string> userIds = sessions.Values.Select(s => s?.User?.Id).Where(id => id != null)!;
 
             var profiles = userIds.Any() ? await GetProfiles(userIds, displayOptions, ctx.CancellationToken) : new Dictionary<string, ProfileDto>();
 
 
-            var results = new Dictionary<string, ProfileDto>();
+            var results = new Dictionary<SessionId, ProfileDto>();
 
             foreach (var session in sessions)
             {
                 var userId = session.Value?.User?.Id;
                 if (userId != null && profiles.TryGetValue(userId, out var profile))
                 {
-                    results.Add(session.Key.ToString(), profile);
+                    results.Add(session.Key, profile);
                 }
             }
             return results;