+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -16,9 +16,12 @@ Changed
 - Updated dependency to Stormancer.Abstractions.Server to 9.2
 - Docker: Support pulling from dockerHub
 - Docker: Use host network mode on Linux
+
 Added
 *****
 - Added IsPersistentServer configuration to set wether the server of a gamesession should be stopped and the scene destroyed when the last player disconnects.
+- Added More analytics
+- Added Agent based game server hosting for better robustness and flexibility.
 
 6.0.2
 -----